Split tests/cross_parity.rs into per-category test files with shared helpers
The monolithic 5,699-line cross_parity.rs (87 test functions) is split into
8 focused test files plus a shared helpers module:

- parity_helpers/mod.rs: shared ICs, assertion fns, Bevy/Sim setup helpers
- bevy_parity_point_mass.rs: point-mass, planetary orbits, timescale, reversal
- bevy_parity_drag.rs: exponential, constant-density, MET atmosphere drag
- bevy_parity_srp.rs: flat-plate SRP, shadow, cannonball, full-stack
- bevy_parity_gravity_torque.rs: gravity torque, external force/torque
- bevy_parity_derived_state.rs: orbital elements, Euler, LVLH, geodetic, beta
- bevy_parity_lighting.rs: earth lighting consistency and geometry tests
- bevy_parity_third_body.rs: 3rd-body, Mars, Mercury, Clementine
- bevy_parity_highfidelity.rs: SH4x4, tidal, polar motion, Gauss-Jackson

All 87 test functions preserved with identical names. Closes #54.
